Illustrate the use of min and max
Adding test which shows how min and max constrain the duration of an animation.

The first element animates for 1 second (because of max).

The second element animates for 2 seconds (because of min).

max=indefinite has no effect, as does min=0. Thus the third element
animates for the requested duration, which is 3 seconds.

min and max are specified at
http://www.w3.org/TR/SMIL3/smil-timing.html#Timing-MinMax

This new test can be viewed at
https://rawgit.com/ewilligers/svg-animation/min-max/examples/compare.html?name=min-max