Illustrate ways of specifying duration
Adds an example that specifies duration in a variety of ways.

Duration values may be specified using clock values
(hour:minute:second.fraction) or using a simple number with units.

If the duration is not supplied, it is assumed to be infinite.

Two of the animations (that specify hour:minute:second) run in Firefox but not Chrome.

The example may currently be viewed at
https://rawgit.com/ewilligers/svg-animation/clock-values/examples/compare.html?name=clock-values